+++ features/bootstrap/ApplicationContext.php
@@ -64,4 +64,12 @@ class ApplicationContext implements Context, SnippetAcceptingContext
     {
         expect($this->application->get($command)->getDescription())->shouldBeLike($text);
     }
+
+    /**
+     * @Then I should see :text in :command command help
+     */
+    public function iShouldSeeInCommandHelp($text, $command)
+    {
+        expect($this->application->get($command)->getHelp())->shouldBeLike($text);
+    }
 }